test(net): make the duplicate-subscribe regression test callback-free
Move the shared transport doubles out of publisher.rs now that both the
publisher and subscriber tests use them, and drop the on-write callback
in favor of a gate the write parks on. The gate is what lets the test
observe the subscription map at the instant the request would reach the
wire, without the transport calling back into the test.

Also assert the recorded bytes decode to exactly one SUBSCRIBE, which
catches a second request encoded onto the same stream, not just a second
stream.
